Refactor by moving deck.gl layers to src/layers/ and loaders to a new src/loaders/
