+++ src/components/cell-tooltip/index.js
@@ -1,6 +1,4 @@
 export { default as CellTooltip } from './CellTooltip';
 export { default as CellTooltipContent } from './CellTooltipContent';
-export { default as CellTooltip1DVertical } from './CellTooltip1DVertical';
-export { default as CellTooltip1DVerticalSubscriber } from './CellTooltip1DVerticalSubscriber';
 export { default as CellTooltip2D } from './CellTooltip2D';
 export { default as CellTooltip2DSubscriber } from './CellTooltip2DSubscriber';
