+++ src/layers/PixelatedBitmapLayer.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import { BitmapLayer } from '@deck.gl/layers';
 import { Texture2D } from '@luma.gl/core';
-import { PIXELATED_TEXTURE_PARAMETERS } from './utils';
+import { PIXELATED_TEXTURE_PARAMETERS } from './bitmap-utils';
 
 
 // These are the same defaultProps as for BitmapLayer.
