+++ src/layers/index.js
@@ -1,3 +1,8 @@
-export { SelectablePolygonLayer } from './SelectablePolygonLayer';
-export { SelectableScatterplotLayer } from './SelectableScatterplotLayer';
-export { getSelectionLayers } from './utils';
+// Selection Layers
+export { default as SelectablePolygonLayer } from './SelectablePolygonLayer';
+export { default as SelectableScatterplotLayer } from './SelectableScatterplotLayer';
+export { getSelectionLayers } from './selection-utils';
+
+// Heatmap Layers
+export { default as HeatmapBitmapLayer } from './HeatmapBitmapLayer';
+export { default as PixelatedBitmapLayer } from './PixelatedBitmapLayer';
