+++ src/layers/selection-utils.js
