+++ src/loaders/index.js
@@ -1,0 +1,3 @@
+export { default as JsonLoader } from './JsonLoader';
+export { default as MatrixZarrLoader } from './MatrixZarrLoader';
+export { default as ClustersJsonAsMatrixZarrLoader } from './ClustersJsonAsMatrixZarrLoader';
